bliss moved to "pddl" namespace
Since we are using bliss with a little bit modified API, we don't want to
allow compiling with the official version. Moreover, we want to allow programs
linking to cpddl to link also to the official version of bliss. So, the local
copy of the bliss library is moved to the pddl namespace.